diverged also JavaSourceProcessor
